Move NamespacedCloudProfile test to project package
The NCP assertions only interact with the NamespacedCloudProfile API
object and need no shoot. Placing them in the project package is a
better semantic fit (NCP is a project-scoped garden resource) and
removes an entire shoot lifecycle (create + hibernate + wakeup + delete)
from the ci-e2e-kind run.
